+++ src/models/RichTextFormatCommand.ts
@@ -5,6 +5,7 @@
  * See the LICENSE.txt file in the project root directory for details.
  */
 import { NoteFontPreference } from "../settings/NoteFontPreference";
+import { NoteFontSize } from "../settings/NoteFontSize";
 
 export enum RichTextFormatCommand {
   BOLD = "bold",
@@ -16,10 +17,17 @@ export enum RichTextFormatCommand {
   BULLET_LIST = "bulletList",
   DASHED_LIST = "dashedList",
   NUMBERED_LIST = "numberedList",
+  FONT_SIZE = "fontSize",
   FONT_FAMILY = "fontFamily"
 }
 
-export type RichTextFormatAction = RichTextFormatCommand | {
-  command: RichTextFormatCommand.FONT_FAMILY;
-  noteFont: NoteFontPreference;
-};
+export type RichTextFormatAction =
+  | RichTextFormatCommand
+  | {
+    command: RichTextFormatCommand.FONT_SIZE;
+    fontSize: NoteFontSize;
+  }
+  | {
+    command: RichTextFormatCommand.FONT_FAMILY;
+    noteFont: NoteFontPreference;
+  };